# openapi.yaml
openapi: 3.0.3
info:
  title: UseZoracle Telegram Bot API
  description: API for managing cryptocurrency accounts, checking balances, and performing transactions using the Coinbase CDP SDK
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://usezoracle-telegrambot.up.railway.app
    description: Production server
...
paths:
...
  /api/positions:
    get:
      summary: Get all positions with filtering options
      description: Retrieve all positions (open, closed, pending) with optional filtering by account name and status
      tags:
        - Positions
      parameters:
        - in: query
          name: accountName
          schema:
            type: string
          description: Filter positions by account name
        - in: query
          name: status
          schema:
            type: string
            enum: [open, closed, pending]
          description: Filter positions by status
      x-code-samples:
        - lang: curl
          source: |
            curl -X GET "http://localhost:3000/api/positions"
        - lang: javascript
          source: |
            fetch('http://localhost:3000/api/positions')
              .then(response => response.json())
              .then(data => console.log(data));
      responses:
        "200":
          description: Positions retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PositionsResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/positions/{positionId}:
    get:
      summary: Get a specific position by ID
      description: Retrieve detailed information about a specific position
      tags:
        - Positions
      parameters:
        - in: path
          name: positionId
          required: true
          schema:
            type: string
          description: The position ID
      x-code-samples:
        - lang: curl
          source: |
            curl -X GET "http://localhost:3000/api/positions/position-id-123"
        - lang: javascript
          source: |
            fetch('http://localhost:3000/api/positions/position-id-123')
              .then(response => response.json())
              .then(data => console.log(data));
      responses:
        "200":
          description: Position retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
        "404":
          description: Position not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/positions/account/{accountName}:
    get:
      summary: Get positions by account name
      description: Retrieve all positions for a specific account
      tags:
        - Positions
      parameters:
        - in: path
          name: accountName
          required: true
          schema:
            type: string
          description: The account name
      x-code-samples:
        - lang: curl
          source: |
            curl -X GET "http://localhost:3000/api/positions/account/Bright"
        - lang: javascript
          source: |
            fetch('http://localhost:3000/api/positions/account/Bright')
              .then(response => response.json())
              .then(data => console.log(data));
      responses:
        "200":
          description: Positions retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PositionsResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/positions/status/{status}:
    get:
      summary: Get positions by status
      description: Retrieve all positions with a specific status (open, closed, pending)
      tags:
        - Positions
      parameters:
        - in: path
          name: status
          required: true
          schema:
            type: string
            enum: [open, closed, pending]
          description: The position status to filter by
      x-code-samples:
        - lang: curl
          source: |
            curl -X GET "http://localhost:3000/api/positions/status/open"
        - lang: javascript
          source: |
            fetch('http://localhost:3000/api/positions/status/open')
              .then(response => response.json())
              .then(data => console.log(data));
      responses:
        "200":
          description: Positions retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Position"
        "400":
          description: Invalid status parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/positions/{positionId}/close:
    post:
      summary: Close a position
      description: Close an open position and calculate PnL
      tags:
        - Positions
      parameters:
        - in: path
          name: positionId
          required: true
          schema:
            type: string
          description: The position ID to close
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - exitTransactionHash
              properties:
                exitTransactionHash:
                  type: string
                  description: The transaction hash of the exit transaction
                  example: "0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef"
      x-code-samples:
        - lang: curl
          source: |
            curl -X POST "http://localhost:3000/api/positions/position-id-123/close" \
              -H "Content-Type: application/json" \
              -d '{"exitTransactionHash": "0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef"}'
        - lang: javascript
          source: |
            fetch('http://localhost:3000/api/positions/position-id-123/close', {
              method: 'POST',
              headers: { 'Content-Type': 'application/json' },
              body: JSON.stringify({
                exitTransactionHash: '0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef'
              })
            })
              .then(response => response.json())
              .then(data => console.log(data));
      responses:
        "200":
          description: Position closed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
        "404":
          description: Position not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Position is not open
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/positions/{positionId}/pending:
    post:
      summary: Set position to pending
      description: Set a position to pending status when creating a limit order that hasn't been triggered yet. This is used for orders that are waiting for the market to reach a specific price (e.g., limit buy/sell orders).
      tags:
        - Positions
      parameters:
        - in: path
          name: positionId
          required: true
          schema:
            type: string
          description: The position ID to set to pending
      x-code-samples:
        - lang: curl
          source: |
            # Set a limit buy order to pending (waiting for price to drop)
            curl -X POST "http://localhost:3000/api/positions/Bright-WETH-limit-buy-1703123456789/pending"
        - lang: javascript
          source: |
            // Set a limit sell order to pending (waiting for price to rise)
            fetch('http://localhost:3000/api/positions/Bright-WETH-limit-sell-1703123456789/pending', {
              method: 'POST',
              headers: { 'Content-Type': 'application/json' }
            })
              .then(response => response.json())
              .then(data => console.log(data));
      responses:
        "200":
          description: Position set to pending successfully (limit order waiting to be triggered)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
        "404":
          description: Position not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/positions/limit-order:
    post:
      summary: Create a limit order
      description: Create a limit order that will trigger when the token reaches a specific price. The order will be set to pending status until the price condition is met.
      tags:
        - Positions
      x-code-samples:
        - lang: curl
          source: |
            # Create a limit buy order (buy when price drops to $2500)
            curl -X POST "http://localhost:3000/api/positions/limit-order" \
              -H "Content-Type: application/json" \
              -d '{
                "accountName": "Bright",
                "tokenAddress": "0x4200000000000000000000000000000000000006",
                "orderType": "buy",
                "targetPrice": "2500.00",
                "amount": "0.1",
                "slippage": "0.5"
              }'
        - lang: javascript
          source: |
            // Create a limit sell order (sell when price rises to $2700)
            fetch('http://localhost:3000/api/positions/limit-order', {
              method: 'POST',
              headers: { 'Content-Type': 'application/json' },
              body: JSON.stringify({
                accountName: 'Bright',
                tokenAddress: '0x4200000000000000000000000000000000000006',
                orderType: 'sell',
                targetPrice: '2700.00',
                amount: '0.1',
                slippage: '0.5'
              })
            })
              .then(response => response.json())
              .then(data => console.log(data));
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - accountName
                - tokenAddress
                - orderType
                - targetPrice
                - amount
              properties:
                accountName:
                  type: string
                  description: The account name to create the limit order for
                  example: "Bright"
                tokenAddress:
                  type: string
                  description: The token contract address
                  example: "0x4200000000000000000000000000000000000006"
                orderType:
                  type: string
                  enum: [buy, sell]
                  description: Type of limit order (buy or sell)
                  example: "buy"
                targetPrice:
                  type: string
                  description: The target price at which the order should trigger
                  example: "2500.00"
                amount:
                  type: string
                  description: The amount of tokens to buy/sell
                  example: "0.1"
                slippage:
                  type: string
                  description: Maximum allowed slippage percentage (optional)
                  example: "0.5"
      responses:
        "200":
          description: Limit order created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
        "400":
          description: Bad request (invalid parameters)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests (rate limited)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/tokens/test/{contractAddress}:
    get:
      summary: Test token metadata
      description: Test token metadata fetching for any token address
      tags:
        - Tokens
      parameters:
        - in: path
          name: contractAddress
          required: true
          schema:
            type: string
          description: Token contract address to test
          example: "0x4200000000000000000000000000000000000006"
      x-code-samples:
        - lang: curl
          source: |
            curl -X GET "http://localhost:3000/api/tokens/test/0x4200000000000000000000000000000000000006"
        - lang: javascript
          source: |
            fetch('http://localhost:3000/api/tokens/test/0x4200000000000000000000000000000000000006')
              .then(response => response.json())
              .then(data => console.log(data));
      responses:
        "200":
          description: Token metadata test completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          description: Bad request (invalid parameters)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          default: false
        error:
          type: string
          description: Error message
        timestamp:
          type: string
          format: date-time
          description: ISO timestamp when the error occurred
...
    Position:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the position
          example: "Bright-0x4200000000000000000000000000000000000006-1703123456789"
        accountName:
          type: string
          description: Name of the account that owns this position
          example: "Bright"
        tokenAddress:
          type: string
          description: Address of the token
          example: "0x4200000000000000000000000000000000000006"
        tokenSymbol:
          type: string
          description: Symbol of the token
          example: "WETH"
        tokenName:
          type: string
          description: Name of the token
          example: "Wrapped Ether"
        amount:
          type: string
          description: Amount of ETH spent on the position
          example: "0.1"
        entryPrice:
          type: string
          description: Price when the position was opened
          example: "2500.123456"
        currentPrice:
          type: string
          description: Current price of the token (for open positions)
          example: "2550.654321"
        pnl:
          type: string
          description: Profit and loss in ETH
          example: "0.002"
        pnlPercentage:
          type: number
          description: Profit and loss percentage
          example: 2.5
        status:
          type: string
          enum: [open, closed, pending]
          description: Current status of the position (open=active trade, closed=completed trade, pending=limit order waiting to be triggered)
          example: "open"
        transactionHash:
          type: string
          description: Hash of the entry transaction
          example: "0xabcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890"
        timestamp:
          type: number
          description: Timestamp when the position was opened
          example: 1703123456
        closedAt:
          type: number
          description: Timestamp when the position was closed (for closed positions)
          example: 1703124000
        exitTransactionHash:
          type: string
          description: Hash of the exit transaction (for closed positions)
          example: "0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef"
    PositionsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Successfully retrieved 5 positions"
        data:
          type: object
          properties:
            open:
              type: array
              items:
                $ref: "#/components/schemas/Position"
              description: List of open positions
            closed:
              type: array
              items:
                $ref: "#/components/schemas/Position"
              description: List of closed positions
            pending:
              type: array
              items:
                $ref: "#/components/schemas/Position"
              description: List of pending positions (limit orders waiting to be triggered)
            summary:
              type: object
              properties:
                totalOpen:
                  type: number
                  description: Total number of open positions
                  example: 3
                totalClosed:
                  type: number
                  description: Total number of closed positions
                  example: 10
                totalPending:
                  type: number
                  description: Total number of pending positions (limit orders)
                  example: 1
                totalPnl:
                  type: string
                  description: Total profit and loss from closed positions
                  example: "0.025"
                totalPnlPercentage:
                  type: number
                  description: Average profit and loss percentage from closed positions
                  example: 5.2
...